+++ src/modules/common/evaluation/EvaluationResultGrid.vue
@@ -144,7 +144,7 @@ onBeforeUnmount(() => {
       v-for="item of props.items"
       :key="item.key"
       type="button"
-      class="btn btn-sm gap-2 justify-start outline-none focus:outline-none text-base"
+      class="btn btn-sm gap-2 justify-start outline-none focus:outline-none text-base font-(family-name:--font-graph)"
       :class="{
         'btn-soft': selected === item.key,
         'btn-ghost': selected !== item.key,